Bug/173/events/munim (#174)
* hard coded request scheme about us (#161)

Co-authored-by: KISA - KAIST International Student Organization <64299863+kisawebkaist@users.noreply.github.com>

* made request scheme change based on development settings (#163)

* XSS fix

Updated 3 files. One is the purify.min.js itself. Another one is the tag_filter.js where the vulnerability actually was. The last one is base.html to include the DOMPurify js file in the source.

* Update .pre-commit-config.yaml

flake8 moved from gitlab to github

* fixed unauthenticated event change security bug

To make the event create and event update action login_required, I added method_decorator to both of the classes. Now it will require the user to be logged in to create or edit an event.

---------

Co-authored-by: Nathan <112761202+n4-t3@users.noreply.github.com>
Co-authored-by: KISA - KAIST International Student Organization <64299863+kisawebkaist@users.noreply.github.com>
Co-authored-by: n4-t3 <nathanmekuria@gmail.com>